Active record association extensions fixed.
They were on belongs_to when they should have been on has_many. The code
still worked, which would suggest that it doesn't need to be there at
all, but it only worked thanks to some freaky undocumented behavious of
associations.

This takes the code coverage back to 100% too, which is always nice.